+++ src/protocol/commands.rs
@@ -6,7 +6,7 @@ use crate::target::Target;
 pub(self) mod prelude {
     pub use super::ParseCommand;
     pub use crate::common::*;
-    pub use crate::protocol::common::hex::{decode_hex, decode_hex_buf, is_hex};
+    pub use crate::protocol::common::hex::{decode_hex, decode_hex_buf, is_hex, HexString};
     pub use crate::protocol::common::thread_id::{
         IdKind, SpecificIdKind, SpecificThreadId, ThreadId,
     };