+++ src/protocol/commands/_vCont.rs
@@ -1,9 +1,13 @@
 use super::prelude::*;
 
-// TODO: instead of parsing lazily when invoked, parse the strings into a
+// TODO?: instead of parsing lazily when invoked, parse the strings into a
 // compressed binary representations that can be stuffed back into the packet
 // buffer, and return an iterator over the binary data that's _guaranteed_ to be
 // valid. This would clean up some of the code in the vCont handler.
+//
+// The interesting part would be to see whether or not the simplified error
+// handing code will compensate for all the new code required to pre-validate
+// the data...
 #[derive(Debug)]
 pub enum vCont<'a> {
     Query,
@@ -15,18 +19,7 @@ impl<'a> ParseCommand<'a> for vCont<'a> {
         let body = buf.into_body();
         match body as &[u8] {
             b"?" => Some(vCont::Query),
-            _ => {
-                for range in body
-                    .split_mut(|b| *b == b'r')
-                    .skip(1)
-                    .flat_map(|s| s.split_mut(|b| *b == b':').take(1))
-                {
-                    let mut range = range.split_mut(|b| *b == b',');
-                    let _ = decode_hex_buf(range.next()?).ok()?;
-                    let _ = decode_hex_buf(range.next()?).ok()?;
-                }
-                Some(vCont::Actions(Actions::new_from_buf(body)))
-            }
+            _ => Some(vCont::Actions(Actions::new_from_buf(body))),
         }
     }
 }
@@ -54,20 +47,14 @@ impl<'a> Actions<'a> {
     pub fn iter(&self) -> impl Iterator<Item = Option<VContAction<'a>>> + '_ {
         match self {
             Actions::Buf(x) => EitherIter::A(x.iter()),
-            Actions::FixedStep(x) => EitherIter::B(
-                Some(Some(VContAction {
-                    kind: VContKind::Step,
-                    thread: Some(*x),
-                }))
-                .into_iter(),
-            ),
-            Actions::FixedCont(x) => EitherIter::C(
-                Some(Some(VContAction {
-                    kind: VContKind::Continue,
-                    thread: Some(*x),
-                }))
-                .into_iter(),
-            ),
+            Actions::FixedStep(x) => EitherIter::B(core::iter::once(Some(VContAction {
+                kind: VContKind::Step,
+                thread: Some(*x),
+            }))),
+            Actions::FixedCont(x) => EitherIter::B(core::iter::once(Some(VContAction {
+                kind: VContKind::Continue,
+                thread: Some(*x),
+            }))),
         }
     }
 }
@@ -103,7 +90,7 @@ pub struct VContAction<'a> {
 pub enum VContKind<'a> {
     Continue,
     ContinueWithSig(u8),
-    RangeStep(&'a [u8], &'a [u8]),
+    RangeStep(HexString<'a>, HexString<'a>),
     Step,
     StepWithSig(u8),
     Stop,
@@ -120,18 +107,8 @@ impl<'a> VContKind<'a> {
             [b'C', sig @ ..] => ContinueWithSig(decode_hex(sig).ok()?),
             [b'S', sig @ ..] => StepWithSig(decode_hex(sig).ok()?),
             [b'r', range @ ..] => {
-                // relies on the fact that start and end were decoded as part of
-                // the initial packet parse.
                 let mut range = range.split(|b| *b == b',');
-                let start = {
-                    let s = range.next()?;
-                    &s[..(s.len() / 2)]
-                };
-                let end = {
-                    let s = range.next()?;
-                    &s[..(s.len() / 2)]
-                };
-                RangeStep(start, end)
+                RangeStep(HexString(range.next()?), HexString(range.next()?))
             }
             _ => return None,
         };
@@ -142,24 +119,21 @@ impl<'a> VContKind<'a> {
 
 /// Helper type to unify iterators that output the same type. Returned as an
 /// opaque type from `Actions::iter()`.
-enum EitherIter<A, B, C> {
+enum EitherIter<A, B> {
     A(A),
     B(B),
-    C(C),
 }
 
-impl<A, B, C, T> Iterator for EitherIter<A, B, C>
+impl<A, B, T> Iterator for EitherIter<A, B>
 where
     A: Iterator<Item = T>,
     B: Iterator<Item = T>,
-    C: Iterator<Item = T>,
 {
     type Item = T;
     fn next(&mut self) -> Option<T> {
         match self {
             EitherIter::A(a) => a.next(),
             EitherIter::B(b) => b.next(),
-            EitherIter::C(b) => b.next(),
         }
     }
 }